Created module, Genomic_Outcome() which is saved in both Modules.R and in Genomic_Outcome_Module_Demo.R
Modified ui.R and server.R such that the "Pre-therapy" tab is produced using the Genomic_Outcome() module but the "Change on therapy" tab is still produced using the scripts diff_UI.R and diff_tab_logic.R for the sake of comparison.

Its worth noting that the lists of subtitles and captions included at the bottom of Modules.R will need to be expanded and/or modified as more modules are produced. In particular: when the Genomic_Outcome() is modified to include Response boxplots they will need their own captions.